pilotos/pilotos-4.0/requirements.txt
# requirements.txt

alembic==1.14.1
colorlog==6.9.0
PyMySQL==1.1.2
python-dotenv==1.0.1
PyYAML==6.0.3
sqlacodegen==2.3.0.post1
